power spectrum

+++ Analysis/main.cpp
@@ -10,10 +10,12 @@ int main( ){
       data[i] = 0.5*sin(2.0*M_PI*i/windowLength)+ 0.5*sin(2.0*M_PI*i/windowLength*10);
     }
 
-    float** powerSpectrum = fft.STFT(data,length,windowLength,1);
+    double** powerSpectrum = fft.STFT(data,length,windowLength,1);
     for (unsigned long i=0; i<windowLength/4.0; i++) {
       cout << powerSpectrum[0][i] << endl;
     }
+
+    //peak* findPeaks = fft.findPeaks(powerSpectrum[0], windowLength/4, 1, 10, 0.6);
 
     delete[] data;
     return 0;
